# mountinfo: implement unescape() using strings.Replacer()