@kolyshkin 2024-07-25
OK, we `unescape` the following fields:
 - (4) root:  root of the mount within the filesystem
 - (5) mount point:  mount point relative to the process's root
 - (9) filesystem type:  name of filesystem of the form "type[.subtype]"
 - (10) mount source:  filesystem specific information or "none"

In general, I would not rely on man pages in such a delicate subject. We have to drink raw and unfiltered C code from the kernel source tree.
 
It looks like mountinfo is shown by [show_mountinfo](https://github.com/torvalds/linux/blob/1722389b0d863056d78287a120a1d6cadb8d4f7b/fs/proc_namespace.c#L135). Let's only look at the fields we're interested in:
 - Root is shown via `show_path` -> `seq_dentry` -> [`mangle_path`](https://github.com/torvalds/linux/blob/1722389b0d863056d78287a120a1d6cadb8d4f7b/fs/seq_file.c#L429), the characters being escaped are `\`, `\n`, `\t` and ` ` (space). Note that some filesystems have their own implementation of `show_path` and may escape different characters, although I can't find any fss doing that in the vanilla kernel source tree.
 - Mount point is shown via `seq_path_root` -> `mangle_path`, same escape characters as above.
 - FS type is shown by [`show_type`](https://github.com/torvalds/linux/blob/1722389b0d863056d78287a120a1d6cadb8d4f7b/fs/proc_namespace.c#L92) -> [`mangle`](https://github.com/torvalds/linux/blob/1722389b0d863056d78287a120a1d6cadb8d4f7b/fs/proc_namespace.c#L87), which escapes ` ` (space), `\t`, `\n`, `\\` and `#`. NOTE the `#`.
 - Mount source (aka "device") is shown either via fs-specific `show_devname` function, or via `mangle` with same rules as for the FS type above (i.e. it also includes #).
 
Looking into fs-specific `show_devname`, I found this:
 - afs: no escaping
 - bcachefs: no escaping
 - btrfs: escapes `" \t\n\\"` (as usual)
 - nfs[4]: escapes `" \t\n\\"` (as usual)
 - cifs: escapes `" \t"` only
 
 Now, these analysis might be wrong and they are definitely incomplete as there are filesystems which source code is not in the vanilla kernel source tree (e.g. aufs), and they may have their own escaping rules.
 
 At the very least, this patch should add `#` to the replace list.