@kolyshkin 2024-07-26
> should we take invalid escape sequences as literal strings? What does Linux itself do
with these?

It looks like kernel uses `string_unescape(..., UNESCAPE_OCTAL)`. Apparently it just replaces any \NNN with a character which octal value is NNN (NNN can be 1 to 3 digits long), with no error reporting (if it can not convert something it just leaves it as is).

Note though that the kernel always outputs 3 octal digits.

I will see if I can simplify my code.